Module molecule_parser::parser

Re-exports

use lexer::Token;

use model::Atom;

use model::Molecule;

use model::add_atoms;

use model::mul_molecule;

Structs

ParseNode
Tree structure that represents a molecule

Enums

Grammar
Grammar nodes supported by our syntax tree

Functions

matching
Find matching bracket

merge_atoms
After flattening the tree we can end up with a result similar to this [("K", 4), ("O", 2), ("N", 2), ("S", 4), ("O", 12)] where oxygen has two entries

parse_atom
Parse exactly one atom and its index (if present)

parse_atoms
Build a ParseNode tree from tokens

parse_close_bracket
Parse a closing bracket for a group of atoms or returns an error.

parse_group
Parse a group of atoms that start with an opening bracket and end with either a closing bracket or with an index after the closing bracket

parse_index
Parse the index of an atom or a group of atoms.

parse_molecule

parse_open_bracket
Parse an opening bracket for a group of atoms or returns an error.